move ensure valid method to shared file
